+++ inst/include/rxode2_control.h
@@ -136,4 +136,5 @@
 #define Rxc_indLinMaxIter 131
 #define Rxc_indLinRichardson 132
 #define Rxc_indLinIteration 133
+#define Rxc_indLinJac 134
 #endif // __rxode2_control_H__