+++ inst/include/rxode2parse_control.h
@@ -135,6 +135,7 @@
 #define Rxc_indLinMaxIter 131
 #define Rxc_indLinRichardson 132
 #define Rxc_indLinIteration 133
+#define Rxc_indLinJac 134
 #define RxMv_params 0
 #define RxMv_lhs 1
 #define RxMv_state 2